BackEnd/requirements.txt
asgiref==3.5.2
astroid==2.11.7
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
coverage==6.4
dill==0.3.5.1
Django==4.0.4
django-cors-headers==3.12.0
django-crontab==0.7.1
django-debug-toolbar==3.5.0
Django-Verify-Email==2.0.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
docxtpl==0.16.4
idna==3.3
isort==5.10.1
Jinja2==3.1.2
lazy-object-proxy==1.7.1
lxml==4.9.1
MarkupSafe==2.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
num2words==0.5.11
passlib==1.7.4
pathspec==0.9.0
Pillow==9.1.1
pipreqs==0.4.11
platformdirs==2.5.2
PyJWT==2.4.0
pylint==2.14.5
PyPDF2==2.10.0
python-docx==0.8.11
python-dotenv==0.20.0
pytz==2022.1
reportlab==3.6.11
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tomlkit==0.11.1
tqdm==4.64.0
tzdata==2022.1
urllib3==1.26.11
wrapt==1.14.1
yarg==0.1.9
psycopg2